test: update UTs
