+++ app/components/UI/Sites/components/SiteRowItemWrapper/SiteRowItemWrapper.test.tsx
@@ -3,6 +3,7 @@ import { render, fireEvent } from '@testing-library/react-native';
 import SiteRowItemWrapper from './SiteRowItemWrapper';
 import type { NavigationProp, ParamListBase } from '@react-navigation/native';
 import type { SiteData } from '../SiteRowItem/SiteRowItem';
+import Routes from '../../../../../constants/navigation/Routes';
 
 // Mock the dependencies
 jest.mock('../../../../Nav/Main/MainNavigator', () => ({
@@ -166,6 +167,19 @@ describe('SiteRowItemWrapper', () => {
     });
   });
 
+  const assertBrowserNavigation = (siteUrl?: string) => {
+    expect(mockNavigation.navigate).toHaveBeenCalledWith(
+      Routes.BROWSER.HOME,
+      expect.objectContaining({
+        screen: Routes.BROWSER.VIEW,
+        params: expect.objectContaining({
+          ...(siteUrl ? { newTabUrl: siteUrl } : {}),
+          fromTrending: true,
+        }),
+      }),
+    );
+  };
+
   describe('Navigation and Press Handling', () => {
     it('should call updateLastTrendingScreen when pressed', () => {
       const { getByTestId } = render(
@@ -182,11 +196,7 @@ describe('SiteRowItemWrapper', () => {
 
       fireEvent.press(getByTestId('site-row-item'));
 
-      expect(mockNavigation.navigate).toHaveBeenCalledWith('TrendingBrowser', {
-        newTabUrl: 'https://example.com',
-        timestamp: 1234567890,
-        fromTrending: true,
-      });
+      assertBrowserNavigation('https://example.com');
       expect(mockNavigation.navigate).toHaveBeenCalledTimes(1);
     });
 
@@ -204,27 +214,7 @@ describe('SiteRowItemWrapper', () => {
 
       fireEvent.press(getByTestId('site-row-item'));
 
-      expect(mockNavigation.navigate).toHaveBeenCalledWith('TrendingBrowser', {
-        newTabUrl: 'https://custom-url.com/page',
-        timestamp: 1234567890,
-        fromTrending: true,
-      });
-    });
-
-    it('should update screen before navigating', () => {
-      const { getByTestId } = render(
-        <SiteRowItemWrapper site={mockSiteData} navigation={mockNavigation} />,
-      );
-
-      const callOrder: string[] = [];
-
-      (mockNavigation.navigate as jest.Mock).mockImplementation(() => {
-        callOrder.push('navigate');
-      });
-
-      fireEvent.press(getByTestId('site-row-item'));
-
-      expect(callOrder).toEqual(['update', 'navigate']);
+      assertBrowserNavigation('https://custom-url.com/page');
     });
 
     it('should handle multiple presses correctly', () => {
@@ -247,10 +237,7 @@ describe('SiteRowItemWrapper', () => {
 
       fireEvent.press(getByTestId('site-row-item'));
 
-      expect(mockNavigation.navigate).toHaveBeenCalledWith(
-        'TrendingBrowser',
-        expect.objectContaining({ fromTrending: true }),
-      );
+      assertBrowserNavigation();
     });
   });
 
@@ -272,11 +259,7 @@ describe('SiteRowItemWrapper', () => {
 
       fireEvent.press(getByTestId('site-row-item'));
 
-      expect(mockNavigation.navigate).toHaveBeenCalledWith('TrendingBrowser', {
-        newTabUrl: 'https://minimal.com',
-        timestamp: 1234567890,
-        fromTrending: true,
-      });
+      assertBrowserNavigation('https://minimal.com');
     });
   });
 });
